Python now print traceback
